shared.mk: Ensure WASI_SYSROOT is exported to the environment
If you don't have WASI_SYSROOT set in the shell and you don't specify
it in the make command, then make will try and find the wasi-sysroot in
some known locations and then set the WASI_SYSROOT accordingly.

However in this case (no shell variable and not specified in the make
command) then when it comes to building the rust stuff, cargo build
fails when trying to generate the bindings as it needs the wasi-sysroot
but doesn't pick it up from the shell environment.

Fix this by exporting WASI_SYSROOT from shared.mk

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>