Rust: Port the Rust Wasm demos to the new 'rusty' wrappers
rusty is a thin wrapper over the generated libunit-wasm bindings to
provide a more native rust like interface.

This gets rid of all the casting and ugly string handling. It massively
reduces the amount of unsafe {} blocks needed, though some still are...

All in all this provides a nice code cleanup.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>