README.md: Update the README and config example
This takes into account the recent porting of the Rust demo applications
to the new 'rusty' wrappers.

The updates relate to the function handler names, the ones provided by
the application now use the uwr_ prefix, the malloc and free handlers
are still using the luw_ prefix as the applications aren't providing
their own versions of those, the ones from libunit-wasm will be used.

This is a useful demonstration of a number of things, but ultimately
that you can call these handlers anything you want.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>