Rust/unit-wasm-sys: Don't require the wasi-sysroot
When I added the luw_http_hdr_get_value() function I needed to include
strings.h, unfortunately I added it to unit-wasm.h instead of
libunit-wasm.c

This then meant that we needed to point bindgen/clang, for generating
the bindings, to the wasi-sysroot for the strings.h file. I guess this
is a system dependant include or something and when we use the
wasm32-wasi target in clang it uses a paired down include search path
list, none of which have strings.h

This then had the knock on effect of requiring to inform cargo build of
the wasi-sysroot path.

Now that we put strings.h in the right place, we don't need the
wasi-sysroot when building the rust stuff.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>