libunit-wasm: Include strings.h in the right file
When I added the luw_http_hdr_get_value() function I needed to include
strings.h, unfortunately I added it to unit-wasm.h instead of
libunit-wasm.c

This had the undesirable effect of requiring the wasi-sysroot when
building the rust stuff for generating the libunit-wasm rust bindings.

By including strings.h in the right file we get rid of that requirement
which a subsequent commit will take care of.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>