Rust/rusty: Add uwr_get_http_content_str()
This function is like uwr_get_http_content() except that it returns a
Rust str.

This is more convenient if you want to operate on the body content
within Rust.

It's worth noting that uwr_get_http_content() returns a non null
terminated buffer which makes it tricky to work with in Rust.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>