libunit-wasm: Add a luw_get_http_total_content_sent() function
This function returns the total amount of content that the Wasm module
has received so far.

This will be used in the rusty API's uwr_get_http_content_str() function
which returns the body content as a string. Due to the body content not
being null-terminated we need to know how much data to use for the
string, for which we are currently using uwr_get_http_content_len(),
which could in some cases return too much if a request is being split
over multiple calls the module.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>