libunit-wasm: Add a luw_req_buf_copy() function
This is analogous to luw_req_buf_append() but rather than appending
request data to the buffer it simply overwrites what's currently there.

This is needed to take advantage of the new ability to receive >4GiB
requests/payloads.

On a new request you would call luw_init_ctx(), luw_set_req_buf() &
open(2).

On subsequent calls to the request_handler (for this same HTTP
request/upload) you would call this new function and then write out the
data to a file.

E.g

  /* In the request_handler */
  if (total_bytes_wrote == 0) {
          luw_init_ctx(&ctx, addr, 0);
          luw_set_req_buf(&ctx, &request_buf, LUW_SRB_NONE);

          fd = open("/var/tmp/large-file.dat", O_CREAT|O_TRUNC|O_WRONLY,
                    0666);
  } else {
          luw_req_buf_copy(&ctx, addr);
  }

  buf = luw_get_http_content(&ctx);
  bytes_wrote = write(fd, buf, luw_get_http_content_sent(&ctx));
  total_bytes_wrote += bytes_wrote;
  if (total_bytes_wrote == luw_get_http_content_len(&ctx)) {
          close(fd);
          total_bytes_wrote = 0;
          luw_http_response_end();
  }

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>